refactor create drawer handlers module
